Fixed bool mapping from c
Former-commit-id: f17db49941b6ebe836e73233740f31ca2e95c77f
Former-commit-id: 1a0d3663d1366b9b18c87a4d3407f6ec34ef8712
Former-commit-id: e00a2a68b2e35646084ba4d5926403fb717ee588

+++ include/symbols/types/LgsStr.h
@@ -24,7 +24,7 @@ public:
         addMethod(&isNotEmptyFunc);
         unpackLength = 1;
     }
-    static uint32_t hashString(const string& str);
+    static size_t hashString(const string& str);
     size_t getSizeBytes() override;
     string getIRName() override;
     Type* getIRType() override;
